# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go + Gin backend for a multivendor eCommerce platform. Auth, vendor onboarding (KYC), product catalog, order splitting, commission, settlement.",
        "title": "Coolmate Multivendor eCommerce API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Coolmate Team",
            "email": "dev@coolmate.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/products/pending-approval": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Admin)"
                ],
                "summary": "List products pending approval (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/products/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Admin)"
                ],
                "summary": "Approve product (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/products/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Admin)"
                ],
                "summary": "Reject product (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "description": "Public product catalog with pagination and text search.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Public)"
                ],
                "summary": "List published products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (max 100, default 10)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Substring search on name/description",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/products/{slug}": {
            "get": {
                "description": "Fetches a single published product by its URL slug.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Public)"
                ],
                "summary": "Get product by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/vendor/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Vendor)"
                ],
                "summary": "List vendor products (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Vendor)"
                ],
                "summary": "Create product (STUB)",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/vendor/products/bulk-import": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Vendor)"
                ],
                "summary": "Bulk import products from CSV (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/vendor/products/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Vendor)"
                ],
                "summary": "Update product (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products (Vendor)"
                ],
                "summary": "Archive product (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_coolmate_ecommerce-backend_internal_utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste: `Bearer \u003caccessToken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}